fix(ci): resolve windows clippy + windows/linux test compile failures
Windows clippy (-D warnings), verified locally via x86_64-pc-windows-gnu:
- remove unused imports (windows-only files) / cfg-gate cross-platform ones
- cfg_attr(windows, allow(dead_code)) on macOS-oriented key-combo helpers
- allow(dead_code) on windows job handle getter
- drop redundant u32 cast; cmp_null -> is_null; fix doc list continuation
- cfg-gate unused_mut where mut is unix-only

Windows/Linux test compile:
- provider.rs: annotate rng.gen::<u8>() (E0283 on windows test build)
- guard the three imessage integration tests with #![cfg(target_os = macos)]
  (the imessage channel module is macOS-only)

+++ src/sandbox/platforms/windows/driver.rs
@@ -10,7 +10,7 @@ use std::path::Path;
 use std::time::Duration;
 
 use async_trait::async_trait;
-use tracing::{debug, warn};
+use tracing::debug;
 
 use crate::sandbox::capabilities::SandboxCapabilities;
 use crate::sandbox::command::{SandboxError, SandboxOutput};
@@ -296,7 +296,7 @@ impl OsSandboxDriverTrait for WindowsSandboxDriver {
                 .stdout(std::process::Stdio::piped())
                 .stderr(std::process::Stdio::piped())
                 .stdin(std::process::Stdio::piped())
-                .creation_flags(CREATE_NEW_PROCESS_GROUP as u32)
+                .creation_flags(CREATE_NEW_PROCESS_GROUP)
                 // Belt-and-braces: if our future is dropped (e.g. upstream
                 // cancellation) the OS terminates the child instead of
                 // leaking it. The job object below provides the same
